requirements.txt
asgiref==3.8.1
blinker==1.9.0
chardet==5.2.0
click==8.1.8
colorama==0.4.6
Django==5.1.4
django-filter==25.1
djangorestframework==3.16.0
django-redis==5.4.0
drf-yasg==1.21.10
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
greenlet==3.2.0
gunicorn==23.0.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
opencv-python==4.11.0.86
packaging==25.0
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2024.2
uritemplate==4.2.0
Werkzeug==3.1.3